+++ src/brave-search/interfaces.ts
@@ -1,0 +1,48 @@
+export interface BraveWeb {
+  web?: {
+    results?: Array<{
+      title: string;
+      description: string;
+      url: string;
+      language?: string;
+      published?: string;
+      rank?: number;
+    }>;
+  };
+  locations?: {
+    results?: Array<{
+      id: string; // Required by API
+      title?: string;
+    }>;
+  };
+}
+
+export interface BraveLocation {
+  id: string;
+  name: string;
+  address: {
+    streetAddress?: string;
+    addressLocality?: string;
+    addressRegion?: string;
+    postalCode?: string;
+  };
+  coordinates?: {
+    latitude: number;
+    longitude: number;
+  };
+  phone?: string;
+  rating?: {
+    ratingValue?: number;
+    ratingCount?: number;
+  };
+  openingHours?: string[];
+  priceRange?: string;
+}
+
+export interface BravePoiResponse {
+  results: BraveLocation[];
+}
+
+export interface BraveDescription {
+  descriptions: {[id: string]: string};
+}
