+++ src/brave-search/tools.ts
@@ -1,0 +1,9 @@
+import { DESCRIPTION as WEB_SEARCH_TOOL } from "./tools/webSearch.js";
+import { DESCRIPTION as IMAGE_SEARCH_TOOL } from "./tools/imageSearch.js";
+import { DESCRIPTION as LOCAL_SEARCH_TOOL } from "./tools/localSearch.js";
+
+export default [
+  WEB_SEARCH_TOOL,
+  IMAGE_SEARCH_TOOL,
+  LOCAL_SEARCH_TOOL,
+];
